chat: use peer contact method in recording
There was an inconsistency between in call and out of call chat
messages. In call chat messages set the ContactMethod to that of
the current account when sending a message to the peer, while out
of call messages set it to the CM of the peer. The side effect of
this was that sending an in call chat message would make a CM with
your own URI (eg: your own RingID) appear in the RecentModel
because sending a new message executes a setLastUsed() on its CM.

In order to resolve this inconsistency and the related RecentModel
bug, we now set the CM to that of the peer when sending an in call
message.

This does not lose any information as the text recording still saves
the account with which the messages were sent.

Tuleap: #259
Change-Id: I39921fd902369628f9ee68ba7ccd5667bf7f7803