Load saved TextRecordings from .json files.
The main goal of this patch is to be able to recover ContactMethods
and conversations which have not been added to the users's contacts
and/or which have no call history. This happens in the case that the
users only sends chat messages to a ContactMethod.

Note that there is a possibility for more than one file to be
associated with the same CM, in the case the sha1 of that CM has
changed (eg: a contact UID has been associated with the CM at some
point). In this case, the most recent recording is used. However,
this is a problem which already exists and is unrelated to this patch.

Tuleap: #270
Change-Id: I5c67be83028af3fc3e1950909574ec947c55c50c