Call: do not use the contact name as the peerName
The name of the contact is not the same as the peerName which comes
from the display name. The display name comes directly fromt the
peer (transfered via SIP info) where as the contact is set by the user.

If the contact is deleted or updated, we do not want their old name
to remain in the call history.

Change-Id: Ibae3967291047d1f53c75ef1a043379f2f293fc9
Tuleap: #307