Update patch set 3
Patch Set 3:

My bad i should explain better in the commit msg when this is ready. The intent is to provide error message to end-user not debug information to us. As such it would only be used for end-user friendly info/error message (e.g called failed because, msg could not be sent, etc...). At this time we have no way of providing feedback to what s happening, this is not user friendly.

Patch-set: 3