Update patch set 3
Patch Set 3:

There is a way, but only the KDE client use it an is technically complementary of your patch. It is the AccountStatusModel. A while back, if you recall, this class had 200 error strings for translation, but Yomguy removed them because they were too scary for the translators. That model is/was the intended "resting place" of errors. Rather than have objects floating around, the errors are just logged into that model (per account) and displayed in popups 

In the KDE client, there is status messages on the canvas itself and notifications when the app isn't focussed (currently broken, but the code is there).

The advantage of this is that it create a log and provide a centralized processing point for such errors.

The AccountStatusModel name suck, but PJSIP do send relevant user visible errors. It is too bad the translation was dropped, the reason call failed now appear in english in all languages.

Patch-set: 3
Reviewer: Gerrit User 11 <11@1e26b4c1-6b8c-47ff-9df7-501b633d3b5a>
Label: Code-Review=0