Update patch set 1
Patch Set 1:

The issue that "JOIN" represent that intent. As half-implemented, JOIN is a multi selection action that appear when you select multiple Calls (or a CM/Person and a Call, it should almost work)

This work using the Stepan "converged" object type API. Anything that expose the Ring roles should work.

For for JOIN, it should detect if the call is already in a conference, then offer "detach" else offer "join".

But the whole multi-selection part is not complete, no client support it, some, while the code was there (I removed some in September because of it was unmaintained) I am quite sure I broke it.

Patch-set: 1