history: Fix regression + interger overflow
While not documented, "0" was used as "unlimited hsitory", but
with the recent change, it became "no history". It may make sense
to use -1 as unlimited, but that would break compatibility for
those who imported sflphone configs into Ring.

Also use "long long" as holder for the seconds count. LLVM
UBSAN asserted on an interget overflow on (32bit) ARMv7.

Change-Id: I7137d7f015fd3b42bd46624fa342a39246f05a59
Tuleap: #103