Update patch set 4
Patch Set 4:

> a false positive

It is not the static analyser, but UNSAN, the equivalent of ASAN. It happened because I set the limit to 99999. As the multiplication is by 87000, even such a "small" value will overflow. I could limit the combobox to 9999, but just using a long long fix the API itself, allowing ridiculous numbers to be used and letting the clients set their own limit rather than having the API itself limited.

Patch-set: 4