Update patch set 2
Patch Set 2:

@Emmanuel,
Maybe you should start this discussion on ring, or ring-dev ML... last time I tested the profiles implementation, the main thing that wasn't clear to me was how they're going to be managed, ie: how do you choose between display name, received profile, and contact... or how do you manage potentially merging profiles. Also, I think that the default profile needs to be somehow marked as such, or maybe even created only in memory... there is no reason to send it over to the peer if all it contains is "Unknown" as the name. Obviously these are issues we can tackle incrementally, but it would be nice to have a discussion about them, I think.

Patch-set: 2
Label: Code-Review=0