Update patch set 1
Patch Set 1: Code-Review-1

(1 comment)

Note that removing the scheme was initially done on purpose to let the daemon pick the most secure one (sip: vs. sips:).

I have nothing against the change itself, but why?

Btw, do not trust protocolHint(). It is a fast detection algorithm and it's **not** accurate. It is why it is called "hint".

Patch-set: 1
Label: Code-Review=-1