Update patch set 1
Patch Set 1:

@Emmanuel: in you case (a ring number has been typed without ring: as header), d_ptr->m_HeaderType is set to NONE, but protocolHint is set to RING. My patch tries to use this hint for setting a correct scheme.
Current version of URI::format() gives the same erronous scheme (i.e. empty scheme) in this case. The problem is not on what to use, but on output of any API that returns schemeNames.
Sure a hint is a hint: that the reason of the "default" case in the switch. I recall the previous version is a direct fallback to sip: header, whatever be the protocol hint. This is a more restrictive and bugged behavior than my patch.
I propose to patch format and use it in fullUri.

Patch-set: 1