Update patch set 2
Patch Set 2:

Yeah the explanation is pretty simple why would i want to have the name used most often displayed on the receiving end when i just changed it ? I don't understand which code could depend on that honestly or i fail to see why this would be a necessity. What added value could that add ? As for the API thing i'll let Guillaume decide.

Patch-set: 2