Update patch set 2
Patch Set 2:

The problem is that in SIP accounts, when you call an extension or something like that you get the wrong caller id and depending if you call or get called, you after get "Unknown" or equivalent if there is a privacy setting or the caller DID failed to be resolved.

If I recall, the Voip.ms wiki have an article about how called id resolution work and why it is unreliable. This is why LRC used a counter by default. It doesn't make much sense for Ring accounts, but for SIP ones, it is still relevant.

Patch-set: 2