Update patch set 2
Patch Set 2:

Lets go, it fix my existing phone (I use a SIP account as mainline phone). This has worked reliably for my for years and I get the right called id.

Don't remove a feature that work for me and many other users and is genuinely useful. Just add a new method for your use case or update preferredName to use some kind of hints and return the right one. In essence, the part of the change that track the last used name, this is good. It is the part that remove a feature I need and use that I have a very big problem with. And also breaking the API for no reason.

Patch-set: 2