[WIP] contactmethod: Make the name concept more flexible
Reality is hard. But, as discussed last week on IRC and gerrit #3885,
LRC handling of ContactMethod names will have to be extended so it
become possible to address various limitations.

The current API was using a dumb instance counter. This fixed an
issue where some calls may miss the DisplayName while other have it.
This is due to how caller name resolution differ between countries.
If a call is rounted outside of its own country, the name resolution
policy and price may change, causing some providers to drop it. If a
call to the same peer end up using another route, then the caller ID
will work. This is mentionned here:

https://wiki.voip.ms/article/Caller_ID

On the other hand, this workaround break the use case where the peer
change her/his display name.

The current API expose too much and wont allow different policies to
be implemented depending on the current context.

**WARNING**: This commit break the API to allow both use case to be
handled internally with the same API.

Change-Id: Ia6a5040fcd6403bd1cc38a8d985632e3f751bf67