Update patch set 1
Patch Set 1:

Ok, so here is a little (promised) help for the profiles.

Currently, there is 2 internal ways to share profiles. One is using the MIME payload during a SIP call and the other is the TrustRequest mechanism.

A trust request is a way for Ring users to say "hello" to each other for the first time. The same concept exist in competing products like Skype and MSN messenger. When the Ring private mode is enabled, a trust request is the only way to reach an user that doesn't already know you. There TrustRequests contains a payload. LRC should check if this is a valid MIME payload, then extract the MIME-Type, then check is the MIME-Type is a Ring profile.

If it is, then the trust request model need to be updated and whatever popup/notification/whatever is used to tell the user there is an incoming trust request. This is potentially quite spammy, so a spam filter should eventually be used, but that's something else and not a problem yet.

The few commits I just pushed close some remaining gaps between certificates (aka, Ring identity) and ContactMethods. This is necessary to have a nice continuity between Peer profiles, peer CM and their trusted identity/certificate.

Patch-set: 1
Reviewer: Gerrit User 11 <11@1e26b4c1-6b8c-47ff-9df7-501b633d3b5a>
Label: Code-Review=0