im: connect text status api
Message status and the message id given by the daemon are now saved
in the TextRecording model. On the start of the client, message status
is checked in case it was updated while the client wasn't running.
Once a message reaches a final state, the message id is updated to be 0
so that it is no longer checked for status updates.

Change-Id: I61a31bc886125764f49af3d1528e7026580398d8
Tuleap: #647