sourcemodel: update when devices are reloaded
This fixes the bug that the wrong (or no) camera is shown
to be selected during a call when the device list is reloaded
(this typically happens if a camera is un/plugged).

Now the SourceModel tracks the selected camera by device id as
well, since a reload can cause the index of the devices to change.

The SourceModel is now also initialized with the active index set
to the default device from the DeviceModel.

Change-Id: I7a05604027917414246b806f367236b87b6a971d
Tuleap: #643