make sure certificate id is properly formatted
The call was using the uri string saved in the history directly as
the certificate id; however this could be badly formatted. This patch
ensures that the uri is explicitly formatted to only contain the
RingID hash and nothing else.

The uris associated with a Person are also explicitly formatted to
only contain the hash.

Change-Id: Id78d1626c182b2485342eada81825481ff330d73
Tuleap: #712