Revert "uri: don't strip 'ring:'"
This reverts commit 6d5614eb0f6fbfb117f0acffd875a33abbf8036d because
it causes issues. For example, now lrc detects that two RingIDs which
are the same, but one is prefixed with "ring:" and the other is not
are not the same.

The proper solution is likely to explicity format the uri each time
it is used in LRC or a client, dependig on the need, using the format()
method.

Change-Id: Ice67006f3c769b3d4f6e33ada2602d162e4cd6f4
Tuleap: #648