Update patch set 1
Patch Set 1:

Sorry, I saw this, but came too late to warn you. This could be fixed in the phonedirectory ::getNumber() method. However, always be careful when you change that code path. It is very fine tuned. Any changes can have cascading effects. It took a lot of iterations to get it to the point it is today (aka, works ~95% of the time).

Patch-set: 1
Label: Code-Review=0