Update patch set 2
Patch Set 2:

Its okay for clients to format but the way the URI is copied here, information will be lost.

If the string originally had 'ring:', we can identify it as a Ring URI, then we strip it, copy the string, and we can only guess based on number of characters.

Patch-set: 2