Update patch set 2
Patch Set 2:

This is little more work, but given you guys seem to use the protocolHints() in more and more places, it may be worth it.

Currently, URI is just an extended QString. This was originally done because all I needed was the .hostname() and .userinfo() methods. They were later cached and a full parser was implemented. By basing on QString, I had 2 big advantages. First or all, I didn't break my own (private, at the time) API so it was less work. I also re-used QString own copy-on-write mechanism so I avoided a great deal of copying.

However, URI is now much more important and will only become more important as time goes by. It may be time to create a "real" class for this with implicitly shared q_ptr. That way, no information would be lost. I would:

* rename URI to URI_
* remove the QString base
* create a global operator= to convert URI_ to QString so the existing code continue to work (maybe add some ways to track this to avoid it being abused, it will be slower).
* typedef QSharedPointer<URI_> URI

With some work, it could solve the dilemma. Normal access will be slower, but less information will be recomputed when the URI are detached. So in the end, it should be about as fast. Another possibility, but that would take even more time, would be to go modern-C++ and implement your own reference tracking and pass the "facade" using std::move. This would be very, very fast, but long to implement. URI are mostly read-only (new ones are created with .format() ), so implementing copy-on-write isn't necessary.

Patch-set: 2
Reviewer: Gerrit User 11 <11@1e26b4c1-6b8c-47ff-9df7-501b633d3b5a>
Label: Code-Review=0