Update patch set 2
Patch Set 2: -Code-Review

@Alex Viau
I've re-looked at the code, and if a string containing "sip:" or "ring:" is passed to the constructor, that info is not lots. the strip() function takes the d_ptr->m_HeaderType reference as a param and sets the SchemeType. So the format() function does not have to call protocolHint() in that case.

Patch-set: 2
Label: Code-Review=0