person: save full RingID to vcard
The vcards generated by Person::toVcard() will now save the full RingID
ie: containing at minimum the scheme and userinfo. This will now happen
even if the original input did not contain the scheme. Eg: if the user
entered a RingID to call without the "ring:" prefix, the resulting vcard
will still be saved with the prefix.

Change-Id: I3fb2a1e95a14803a15af2999cbd6dc9a4db51841
Tuleap: #648