Update patch set 1
Patch Set 1: Code-Review-1

Hmm... so I guess we potentially want different behaviour for RING vs not RING accounts... I guess we want some kind of "helper" function which returns the uri with a scheme in the case that its either a RingID, or that the scheme was specified in the original uri, but not otherwise.

I will test this.

Patch-set: 1
Label: Code-Review=-1