Update patch set 1
Patch Set 1:

> What are the consequences of not doing anything?

See the attached ticket... we want to save the "ring:..." in the vcard, so its clear that its a ringId and not some other hash... I guess the other option is to use a custom field in the vcard, instead of a phone number... but some lesser contact managers (eg: gnome-contacts) don't support custom fields.

I'll likely avoid adding the helper for now, and simply perform the check in vcardutils.

Patch-set: 1