profile: Partially fix multiple profiles
This commit implements some logic necessary to assign a profile per account
rather than a global one. This will require more work to add both a global
profile mode and a per account one.

A profile is now assigned to an account at all times to avoid any account
belonging to multiple profiles at once. This should make sense in itself
as it avoids a potential logical cardinality issue.

Again, it is not intended to break the Gnome/OS X/Windows clients way of
handling profiles, but make other models possible to implement.

Change-Id: Ibd675210dc2939244454cee338cf9cda263aa2ae