Update patch set 11
Patch Set 11:

> Should I have also implemented RemoveRow?

No, its just a helper that calls RemoveRows.

> I don't know what is the use we have for the QModelIndex in that case, so I just ignored it.

It specifies the parent, but since this model has only a depth of one, the parent of all items should always be the root (ie: and invlaid QModelIndex). The more correct way would be to actually make sure that parent.isValid() == false in the beginning, I think

Patch-set: 11
Label: Code-Review=0