Ensure bootstrapModel is not out if sync with account hostname
The bootstrapModel is constructed from the hostname, setting the
hostname after account creation causes it to fall out of sync. This is
only relevant for Ring accounts.

Modified the reset() method so that it does not assume one existing
bootstrap server, preventing a segfault when reseting an empty
bootstrapModel.

Change-Id: I9942f3d8fd344f9cc4d76556aa063264abb34ac4
Tuleap: #733