Ensure bootstrapModel is not out if sync with account hostname
The bootstrapModel is constructed from the hostname, setting the
hostname after creating the boostrapModel causes the bootstrapModel
to fall out of sync with the hostname. This is only relevant for Ring
accounts.

This patch implements a state machine to manage the life cycle of the
bootstrapModel and ensure it is in sync with the account hostname

The reset() method was modified so that it does not assume one existing
bootstrap server, preventing a segfault when reseting an empty
bootstrapModel.

Change-Id: I9942f3d8fd344f9cc4d76556aa063264abb34ac4