Remove duplicated D-Bus interface files
Now that D-Bus interface files are installed by the Ring daemon,
there is no need to duplicate them in this repository.

This patch introduces a new cmake flag: -DRING_XML_INTERFACES_DIR,
which can be used to specify the directory containing the D-Bus
interface files.

If the flag is not specified, we look for the following directories,
in order:
 - ${RING_BUILD_DIR}/../bin/dbus
   This should work in most cases, where we are building against the
   project directory.

 - ${CMAKE_INSTALL_PREFIX}/share/dbus-1/interfaces
   This is where D-Bus interface files should be installed, as per the
   freedesktop website[1].

1. https://dbus.freedesktop.org/doc/dbus-api-design.html#interface-files

Change-Id: I8e8476c1276a1dbb0af2197a2d2d8efff2724f4e
Tuleap: #688