Update patch set 12
Patch Set 12:

> Sorry for being lazy and not reading the daemon code, but what is this about? What are the returned values and when is this called/necessary?

Read the daemon code :P

This will create an ICE transport in the daemon and return if it was successful. The idea is for the user to be able to verify that the account is able to make a call at the ICE level (since a call can fail for other reasons as well). Primarily, so that when they enter new TURN credentials, they can click a button and make sure it actually connects to the TURN server.

Patch-set: 12
Label: Code-Review=0