Update patch set 12
Patch Set 12:

Ok, then this should probably fall into the Credential object scope. May I suggest to add a parameter to the daemon function to be able to test a specific set of credentials (STUN (not implemented yet?) and TURN)?

This thing is, if you want to add a "check connectivity" button in the client, you probably don't want to apply the whole Account "transaction", but only check a set of credentials without impacting the whole account (and CredentialModel) state.

So, I suggest:

* Move this into ::Credential
* Allow the method to be called without having to perform a `account << Account::SAVE` because it they are different concepts that shouldn't depend on each other.
* If I may, please also turn this into an ASYNC transaction. If it is blocking the UI/MainThread for longer than half a second, then it has to be async. Introducing new blocking methods now isn't really something I will be happy with. It create a bad user experience and is a bad design in general. Adding a dring signal and returning a transation ID takes longer, but not that much. Then on LRC side, a nice little state machine wont be much trouble to implement.

Patch-set: 12