Update patch set 19
Patch Set 19: Code-Review-1

(3 comments)

This is looking pretty good in terms of the API of SmartInfoHub, however there is still some work to do in the code itself.

There are a few comments that both Alexandre and I made that you have not responded to yet. I suggest that you look at all of the comments that have been made in the code, and if you agree, you should fix it (if you haven't yet) and then mark it as "done", if you don't agree or don't understand the comment, you should reply to it directly. This way you can make sure not to miss anything, and we know whether or not you agree/disagree or fixed the issue.

Patch-set: 19
Label: Code-Review=-1