Update patch set 8
Patch Set 8:

> Why? The code is quite nice

Replaced by https://gerrit-ring.savoirfairelinux.com/#/c/4753/.

The new changeset is against a wip/multi-device branch.

In the last iteration meeting, we decided to move changesets that cannot be merged right now to master to wip branches. When we decide to roll-out multi-device, we will merge the wip branches to master.

Patch-set: 8