Update patch set 44
Patch Set 44: Code-Review+2

> Would it make sense to move all the variables into their respective classes

Do you mean that, for example, instead of returning the codec as a string, you return the codec object in the getter? If so, I think this should only be done if at some point we need to compare the values/object somewhere in some logic... as is, this is only used for debugging/display purposes, so it seems to me like it would add a lot of logic (and thus code and potential bugs) for not a lot of benefit.

> _or_ turn that into a TableModel?
What would each row be? The callID? Right now, I think the signal with the info is only emitted for the active call, so only one call's info gets updated at a time.

I think we need to merge as is for now, especially since it doesn't really touch any existing code... we can improve iteratively afterwards based on how this will get used in the future.

Patch-set: 44
Reviewer: Gerrit User 23 <23@1e26b4c1-6b8c-47ff-9df7-501b633d3b5a>
Label: Code-Review=+2