improve RecentModel::getIndex(ContactMethod*) behaviour
ContactMethods which are associated with a Person are not contained
in the RecentModel as items. Thus we amend the getIndex function to
also check if the given ContactMethod is associated with a Person
which is an item in the model, instead of just returning an invalid
QModelIndex in this case.

Change-Id: I996d08bb7d9a550bd3f027348b350b3b0bc631b9
Tuleap: #956