add clear method to delete chat history
This will delete all the .json chat files in the text dir. It also
clears all the TextRecording models, though the models themselves
remain to prevent nullptr errors. Also this does not remove the
ContactMethods which were generated from the chat history, the user
will have to re-launch ring for this to take effect.

Change-Id: Ifc3cf6289d19512bfc9acb5ea1c9de7b159f4473
Tuleap: #925