IM: correctly emit dataChanged for both text models
This commit updates the accountMessageStatusChanged method to emit
dataChanged signals on both the ImModel and TextMessagesModel. These
signals are used for delivery reports by clients.

We also expose the message id from the daemon. It can be used by clients
to identify messages and it is also used to match and find the
QModelIndex of the message we received an update for.

Tuleap: #1073
Change-Id: Iee9529e374086a268528d06e91c9f5fe0e61811d