IM: correctly emit dataChanged on the right QModelIndex
This commit updates the accountMessageStatusChanged method to emit
dataChanged signals on the right QModelIndex.

This is done by saving the row of the QModelIndex in the message node.
When and if re-ordering is implemented, this number will have to be
updated.

We also expose the message id from the daemon. It can be used by clients
to uniquely identify messages. This will be useful when re-ordering is
implemented.

Tuleap: #1073
Change-Id: Iee9529e374086a268528d06e91c9f5fe0e61811d