invoke ContactMethod::addCall() even if the call fails
If the call immediately failed on creation (empty call id returned
by the daemon) the call was not being added to the CM nor the Person
because addCall() was never being called. This created an
inconsistency since the call IS added to the CallModel. Such calls
should either not be added to the CallModel, or else they should be
added to the CM and Person as well. We choose the latter solution
as this is more informative to the user. Otherwise the user has no
way of knowing if the call failed immediately or was just never even
attempted.

Thus this patch ensures ContactMethod::addCall() is called for all
attempted calls. The PhoneDirectoryModel is ammended to not keep
track of statistics for added calls which are in a failed state.

Change-Id: Ie9d2c22a8a9b7fe84d4d02b3887901ac8974130f
Tuleap: #1093