update the lastUsed of ContactMethod during outgoing call
When starting an outgoing call, the lastUsed timestamp of the
related contact method was not being updated due to a regression
caused by commit 596f988d2e29dae3bf9e79d6484c977e6180df2d

This resulted in that ContactMethod not moving to the top of the
RecentModel list as it should.

This patch fixes this by ensuring that the start timestamp of the
call is set before its added to its ContactMethod.

This means that even calls which fail immediately will have a start
timestamp. However, this is currently the case for calls which
correctly initialize but then fail later. Thus this behaviour is
more consistent.

Change-Id: I2de74e059bb174d192993f6a508e3d66c75e0166
Tuleap: #1110