CallModel: revert disconnecting from removed call
Commit 6b78703012622a4c26c651c8752e3166026e3312 introduced a change
which would disconnect the CallModel from all signals of a removed
call. However, this is done in CallModelPrivate::removeInternal
which is not always used to remove a call permanatly, sometimes its
used to move a call within the CallModel, for example in the case
of conferences.

Another regression which was introduced by that change was that
calls would no longer be properly removed from the RecentModel
when a conference call was over.

Reverting this change fixes these two issues.

Change-Id: I889e083017da3c816a7b7b7314d73831cd03b991
Tuleap: #1105