Prevent QVector index out of bounds crash
Its not clear why this code exists in the first palce, but its
trying to access a QVector with an index of -1. It seems that the
code did not enter here until recently so its not clear if it should
exist at all, but this should at least prevent the crash.

Change-Id: Id80497b800b9f9f111492336f436c63eafe4e2a5
Tuleap: #1182