Perform name lookup on new RingIDs
When a new ContactMethod is created, if its URI is a RingID, we
perform a name lookup to the Ring username (if it exists). We also
do this in the case that the account of that CM is changed since
each account could have a different name server configured.

Change-Id: I34122ac92fcaad0006c28f3ec88c787f01ba6000
Tuleap: #1169