presence: load tracked buddies
Load track buddies when creating an account. We do this so that the
state of contactmethods is as up to date as possible.

Otherwise, restarting the client would show the contacts as offline
because we wait for callbacks of the daemon to set the presence status.

This loads the status of all currently tracked buddies.

Tuleap: #1379
Change-Id: I047900bcbec9196f66d0d425e573f64090e3eba7