Update patch set 1
Patch Set 1:

@Stepan Salenikovich
I'm working to modify a lot, how the contacts are managed in lrc. I found several bugs in lrc. just do a grep over the project and you'll see where the constructor of Person is called without valid uid. The current situation lead to at least one bug, and iirc to two. The one i remember is the about the way the filename of the vcf are generated. Without uid the vcf has no name, leading to a file called ".vcf" . I try to add less changmeent in the code that's why I think this solution is better. The question I would like to ask is : in which case do you would like to have no uid at all ? because this fix ensure an uid, but you can overwrite it. So far, the value itself has no importance as long as it's unique. Guillaume got a point about some fast call, I will remove the time function by some random function.

Patch-set: 1
Label: Code-Review=0