Update patch set 3
Patch Set 3:

Yeah, we could just remove it from the constructor. I think the main point is that uid's are only relevant to the Collections when storing contacts (Persons), so we should somehow work towards the uid only being used by the Collections and ideally not in any Person logic itself.

> does it make sense to let someone to create a Person by calling directly the constructor ? should we not delegate the creation of Person instances to PersonModel ?

Not sure. Right now its the collections which call the constructor and even the clients when creating a new contact. We could rework the system to be more like ContactMethods/PhoneDirectoryModel. ie: have everyone pass a set of parameters to PersonModel and have it construct the Person. But I'm not sure what the exact benefits of all that work will be. Maybe if we want to force some kind of automatic de-deuplication of contacts. ie: the Person model would be in charge of somehow detecting duplicate contacts. Anyways, this is sort of another subject.

Patch-set: 3