Update patch set 3
Patch Set 3:

There is always const_cast<>() ... not idea, but thats whats done in FallbackPersonBackendEditor::save().

Of course, the other solution is like you said to change the base class. I'm open to either one... it is rather silly to declare a function as const, and then in every definition of that function use a const_cast...

Patch-set: 3