Update patch set 1
Patch Set 1:

FYI, data and layout changed signals, as well as many others come from the QAbstractItem model. They are needed by the views (ie: the clients) to know when the model has changed. So you would have to check the clients to see if they are used. Though you wouldn't see them in the Qt (windows) client, since its the Qt code which connects to them.

Patch-set: 1
Label: Code-Review=0