Update patch set 1
Patch Set 1: Code-Review-1

I think for all of these models (CredentialModel, BootstrapModel, etc) we set the Account object passed to their constructor as their parent. ie: the Account is passed to the constructor of the QAbstractItem model, which is passed to the QObject constructor. This means that when the Account object is destroyed, it will try to destroy its children QObjects, as per the doc:
http://doc.qt.io/qt-5/qobject.html#QObject

Again, as per the doc, we can think of the Account object as the owner of these models due to this mechanism.

If we now place these pointers into a unique_ptr, then this provides a second mechanism by which the Account object will be their owner.

IMHO, only one of these should be used. ie: we should either not set the Account as the parent via the constructor or we shouldn't use the unique_ptr.

Patch-set: 1
Reviewer: Gerrit User 23 <23@1e26b4c1-6b8c-47ff-9df7-501b633d3b5a>
Label: Code-Review=-1