Update patch set 1
Patch Set 1:

>right! That's a good reason to dropt Qt support with such implicit rules.

Mmm, well, I'm not sure if its an "implicit" rule. We explicitly pass the Account pointer to set it as the parent of the these other models in lrc. This seems typical of UI frameworks (GTK+ does a similar thing) since in most cases the object correspond to widgets, and so in most cases you want the parent to destroy its children widgets when it gets destroyed. Maybe we simply don't want to do it in this case since we're dealing with models and not widgets. Its perfectly valid to pass a nullptr as the parent argument.

> And in this case, the 2 delete calls into ~Account are bugs, exact?

I'm not sure. I quickly looked into the QObject code, but its complicated. Its possible that when, for example, we explicitly delete the BootstrapModel, it will notify its parent, and so Account will not try to delete it a second time when it gets destroyed.

Patch-set: 1