Update patch set 5
Patch Set 5: Code-Review-1

(8 comments)

the code itself is ... hmm ... "ok". Not your fault but looks like coded by a "C" people.
For sure its due to all others lines around and global design.
My biggest problem is the leak of documentation and the variable name obfuscating everywhere. Don't give better chance to understand of check if it's good and if nothing has been forgotten.

So, for me only typo has to be fixed, the code "works"... but not really checkable without to be the author of the design.

Patch-set: 5
Label: Code-Review=-1