Update patch set 2
Patch Set 2:

> Not sure, Registration is an option, a Ring account without registred 
> name is a valid account. But now, for a Ring account, do we need to 
> use DisplayName or Alias?

It's not displayed in either the Gnome or macOS client, as the KDE (and legacy Windows Qt client) client use the field availability to decide if it's in the account setting page or not, I figured out this patch was necessary. Feel free to disagree, you know the right answer better than me on this, all I did was to update the list to reflect the current Gnome client GUI.

If the display name still has a reason to be, then at least we should agree to have it in the other clients too. One way or another, as long as it's consistent. Anyway, neither the Gnome or macOS client use this list, so the change have a very low impact on Ring.

Patch-set: 2
Label: Code-Review=0