Update patch set 7
Patch Set 7: Code-Review-2

In general, hostname is a crutial field of Ring accounts.
So saying is not "available" (remain to explicit what we are saying under this term) is a non-sens here.
If I read the documentation of the modified method, it's usage is to indicate if a field has a "sens" for a specific account.
As such, I don't want to submit this patch.

But, I recommend clients to use bootstrap model if they don't want to handle a complex field.
This method must remains a hint for clients, but not enforce them what to display.

Patch-set: 7
Reviewer: Gerrit User 3 <3@1e26b4c1-6b8c-47ff-9df7-501b633d3b5a>
Label: Code-Review=-2