Update patch set 7
Patch Set 7:

I think you are unaware of an important detail here. It is true that the hostname field is important, however, this is not related to this patch. The hostname of ring account is a CSV array of bootstrap nodes. LRC exposes it as a list (called the bootstrap model). This is how the Gnome and macOS client display it in the GUI.

This patch exist only to synchronize the behavior of the Gnome and macOS clients with the KDE client. It only standardize what is already done in the other clients, nothing else. If you refuse to merge it because the legacy Windows client doesn't display the list yet, I can take care of it.

Patch-set: 7
Reviewer: Gerrit User 11 <11@1e26b4c1-6b8c-47ff-9df7-501b633d3b5a>
Label: Code-Review=0