cmake: Support using Ring-LRC as a "cmake external project"
This is the CMake equivalent of the contrib system. You can add
external projects as git urls.

The commit itself only cleanup the build files to make sure it
doesn't cause any error due to confusing the project path with
the current module path.

It also allow FindRing to pass if the XML path is defined.

Another cmake module allows the XML to be fetched from Github
without having to compile dring. This simplify the CI
integration for embedded Linux.

Change-Id: I56f0fac87c0c75706bae2c5dfcafcaad2144e253