Update patch set 8
Patch Set 8:

> (1 comment)

As you mention "if the certificate is a valid Ring one" the problem is here.
When only a RingID is available, the daemon doesn't return a certificate. So an empty certificate is attached to CM, and we can't progress further. Daemon API sendTrustRequest accepts only a RingID as input, so I've made a change relative to this fact.

Patch-set: 8