Update patch set 1
Patch Set 1:

> wagaf point out that the CA still exists.

True. A CA is always mandatory.
In case of Ring account created by Ring, this
CA is self-signed and used to sign the RingID key, this one used
to sign the DeviceID key.

 > However it doesn't seem
 > to be visible in getaccountdetails after creating a Ring account.
 > 
 > in d-feet:
 > 
 > "TLS.certificateListFile" = ""
 > 
 > So either:
 > 
 > 1) It has no place in the account details. In that case this patch
 > is correct
 > 2) It's a daemon bug, in that case, the current behavior is correct
 > and the bug need to be fixed in the daemon.
 > 
 > Either way, it doesn't matter much. As log as the "secure" icon
 > gets green again, that will do. Given the CA is internal to dring,
 > I think '1)' is good enough, but feel free to disagree.

TLS.certificateListFile is non-empty only if a specific CA certificate
is given as a file.
In case of device certificate, only TLS.certificatFile is filled and the former remains empty as the CA is already available into the later.

Patch-set: 1