Update patch set 1
Patch Set 1: Code-Review-2

As I've previously said, CA is mandatory.
It must be found into the TLS.certificateFile if TLS.certificateListFile is empty.
So relaxing the security evaluation when CA is not found is just an error.

If you get an error message due to empty CA file, but a regular certificate with included CA is present, it's a bug into LRC, but it's not here.

Patch-set: 1
Reviewer: Gerrit User 3 <3@1e26b4c1-6b8c-47ff-9df7-501b633d3b5a>
Label: Code-Review=-2