Update patch set 1
Patch Set 1:

> If you get an error message due to empty CA file, but a regular certificate with included CA is present, it's a bug into LRC, but it's not here.

Actually, this just confirm this patch. In older versions, the CA was in its own file, now it's built in. This error is due because LRC still expect it in a CA list file in `certificateListFile`. This is now wrong and it the source of the error.

Patch-set: 1
Label: Code-Review=0